refactor: extract logic from e2e tests to reuse it for integrations tests
E2E tests run an independent application process with docker, and then they are executed against that public service. It's only one instance.

Integrations tests will use an independent process for each test, running the API on a different port. The instance SUT is launched by the tests. We need to refactor the application bootstrapping to be able lo launch an app instance from tests.

This refactor allows us to reuse the code from E2E tests in future integration tests.
