+++ tests/common/connection_info.rs
