+++ tests/common/contexts/category/forms.rs
@@ -1,0 +1,9 @@
+use serde::Serialize;
+
+#[derive(Serialize)]
+pub struct AddCategoryForm {
+    pub name: String,
+    pub icon: Option<String>,
+}
+
+pub type DeleteCategoryForm = AddCategoryForm;
