+++ tests/common/contexts/category/mod.rs
@@ -1,0 +1,3 @@
+pub mod fixtures;
+pub mod forms;
+pub mod responses;
