+++ tests/common/contexts/mod.rs
@@ -1,0 +1,6 @@
+pub mod about;
+pub mod category;
+pub mod root;
+pub mod settings;
+pub mod torrent;
+pub mod user;
