+++ tests/common/contexts/settings/form.rs
@@ -1,0 +1,3 @@
+use super::Settings;
+
+pub type UpdateSettingsForm = Settings;
