+++ tests/common/contexts/torrent/fixtures.rs
@@ -6,11 +6,9 @@ use tempfile::{tempdir, TempDir};
 use uuid::Uuid;
 
 use super::file::{create_torrent, parse_torrent, TorrentFileInfo};
-use super::requests::{BinaryFile, UploadTorrentMultipartForm};
-use super::responses::{Id, UploadedTorrentResponse};
-use crate::e2e::contexts::category::fixtures::software_predefined_category_name;
-use crate::e2e::contexts::user::LoggedInUserData;
-use crate::e2e::environment::TestEnv;
+use super::forms::{BinaryFile, UploadTorrentMultipartForm};
+use super::responses::Id;
+use crate::common::contexts::category::fixtures::software_predefined_category_name;
 
 /// Information about a torrent that is going to added to the index.
 #[derive(Clone)]
@@ -51,26 +49,6 @@ impl TorrentListedInIndex {
             torrent_file: torrent_to_index.torrent_file,
         }
     }
-}
-
-/// Add a new random torrent to the index
-pub async fn upload_random_torrent_to_index(uploader: &LoggedInUserData) -> (TestTorrent, TorrentListedInIndex) {
-    let random_torrent = random_torrent();
-    let indexed_torrent = upload_torrent(uploader, &random_torrent.index_info).await;
-    (random_torrent, indexed_torrent)
-}
-
-/// Upload a torrent to the index
-pub async fn upload_torrent(uploader: &LoggedInUserData, torrent: &TorrentIndexInfo) -> TorrentListedInIndex {
-    let client = TestEnv::default().authenticated_client(&uploader.token);
-
-    let form: UploadTorrentMultipartForm = torrent.clone().into();
-
-    let response = client.upload_torrent(form.into()).await;
-
-    let res: UploadedTorrentResponse = serde_json::from_str(&response.body).unwrap();
-
-    TorrentListedInIndex::from(torrent.clone(), res.data.torrent_id)
 }
 
 #[derive(Clone)]
