+++ tests/common/contexts/torrent/forms.rs
@@ -1,10 +1,15 @@
 use std::fs;
 use std::path::Path;
 
-use reqwest::multipart::Form;
 use serde::{Deserialize, Serialize};
 
-pub type TorrentId = i64;
+#[derive(Deserialize, Serialize)]
+pub struct UpdateTorrentFrom {
+    pub title: Option<String>,
+    pub description: Option<String>,
+}
+
+use reqwest::multipart::Form;
 
 pub struct UploadTorrentMultipartForm {
     pub title: String,
@@ -42,10 +47,4 @@ impl From<UploadTorrentMultipartForm> for Form {
                     .unwrap(),
             )
     }
-}
-
-#[derive(Deserialize, Serialize)]
-pub struct UpdateTorrentFrom {
-    pub title: Option<String>,
-    pub description: Option<String>,
 }
