+++ tests/common/contexts/torrent/requests.rs
@@ -1,0 +1,1 @@
+pub type TorrentId = i64;
