+++ tests/e2e/contexts/about/mod.rs
@@ -1,2 +1,1 @@
 pub mod contract;
-pub mod steps;
