+++ tests/e2e/contexts/root/mod.rs
@@ -1,2 +1,1 @@
 pub mod contract;
-pub mod steps;
