+++ tests/e2e/contexts/settings/contract.rs
@@ -1,91 +1,8 @@
-use serde::{Deserialize, Serialize};
-
-use crate::e2e::contexts::user::fixtures::logged_in_admin;
+use crate::common::contexts::settings::form::UpdateSettingsForm;
+use crate::common::contexts::settings::responses::{AllSettingsResponse, Public, PublicSettingsResponse, SiteNameResponse};
+use crate::common::contexts::settings::{Auth, Database, Mail, Net, Settings, Tracker, Website};
+use crate::e2e::contexts::user::steps::logged_in_admin;
 use crate::e2e::environment::TestEnv;
-
-// Request data
-
-pub type UpdateSettingsForm = Settings;
-
-// Response data
-
-#[derive(Deserialize)]
-pub struct AllSettingsResponse {
-    pub data: Settings,
-}
-
-#[derive(Deserialize, Serialize, PartialEq, Debug)]
-pub struct Settings {
-    pub website: Website,
-    pub tracker: Tracker,
-    pub net: Net,
-    pub auth: Auth,
-    pub database: Database,
-    pub mail: Mail,
-}
-
-#[derive(Deserialize, Serialize, PartialEq, Debug)]
-pub struct Website {
-    pub name: String,
-}
-
-#[derive(Deserialize, Serialize, PartialEq, Debug)]
-pub struct Tracker {
-    pub url: String,
-    pub mode: String,
-    pub api_url: String,
-    pub token: String,
-    pub token_valid_seconds: u64,
-}
-
-#[derive(Deserialize, Serialize, PartialEq, Debug)]
-pub struct Net {
-    pub port: u64,
-    pub base_url: Option<String>,
-}
-
-#[derive(Deserialize, Serialize, PartialEq, Debug)]
-pub struct Auth {
-    pub email_on_signup: String,
-    pub min_password_length: u64,
-    pub max_password_length: u64,
-    pub secret_key: String,
-}
-
-#[derive(Deserialize, Serialize, PartialEq, Debug)]
-pub struct Database {
-    pub connect_url: String,
-    pub torrent_info_update_interval: u64,
-}
-
-#[derive(Deserialize, Serialize, PartialEq, Debug)]
-pub struct Mail {
-    pub email_verification_enabled: bool,
-    pub from: String,
-    pub reply_to: String,
-    pub username: String,
-    pub password: String,
-    pub server: String,
-    pub port: u64,
-}
-
-#[derive(Deserialize)]
-pub struct PublicSettingsResponse {
-    pub data: Public,
-}
-
-#[derive(Deserialize, PartialEq, Debug)]
-pub struct Public {
-    pub website_name: String,
-    pub tracker_url: String,
-    pub tracker_mode: String,
-    pub email_on_signup: String,
-}
-
-#[derive(Deserialize)]
-pub struct SiteNameResponse {
-    pub data: String,
-}
 
 #[tokio::test]
 #[cfg_attr(not(feature = "e2e-tests"), ignore)]
