+++ tests/e2e/contexts/settings/mod.rs
@@ -1,2 +1,1 @@
 pub mod contract;
-pub mod steps;
