+++ tests/integration/contexts/about.rs
@@ -1,0 +1,5 @@
+#[tokio::test]
+#[ignore]
+async fn it_should_load_the_about_page_with_information_about_the_api() {
+    // todo: launch isolated API server for this test
+}
